4.21

+++ controller/blogs.js
@@ -57,6 +57,27 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" })
+  }
+
+  if (!request.token) {
+    return response.status(401).json({ error: "token required" })
+  }
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response
+      .status(403)
+      .json({ error: "not authorized to delete this blog" })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
